Recap on async await patterns and event loops
