--- asyncAwait.js
@@ -1,51 +1,0 @@
-const { readFile, writeFile } = require("fs").promises;
-// const util = require("util");
-
-/*
-! PROMISES
-*/
-
-// const getText = (path) => {
-//   return new Promise((resolve, reject) => {
-//     readFile(path, "utf8", (err, data) => {
-//       if (err) {
-//         reject(err);
-//       } else {
-//         resolve(data);
-//       }
-//     });
-//   });
-// };
-
-// getText("./content/first.txt")
-//   .then((result) => {
-//     console.log(result);
-//   })
-//   .catch((err) => {
-//     console.log(err);
-//   });
-
-/*
-! Refactor to async await
-? I am mind blown by this code to avoid callback hell!!
-*/
-
-// const readFilePromise = util.promisify(readFile);
-// const writeFilePromise = util.promisify(writeFile);
-
-const start = async () => {
-  try {
-    const first = await readFile("./content/first.txt", "utf8");
-    const second = await readFile("./content/second.txt", "utf8");
-    await writeFile(
-      "./content/result-mind-grenade.txt",
-      `THIS IS AWESOME:\n${first}\n${second}`,
-      { flag: "a" }
-    );
-    console.log(first, second);
-  } catch (error) {
-    console.log(error);
-  }
-};
-
-start();
