+++ content/result-mind-grenade.txt
@@ -1,11 +1,5 @@
-THIS IS AWESOME:
-Hello this is the first text file
-Hello this is the second text file!THIS IS AWESOME:
-Hello this is the first text file
-Hello this is the second text file!THIS IS AWESOME:
-Hello this is the first text file
-Hello this is the second text file!THIS IS AWESOME:
-Hello this is the first text file
-Hello this is the second text file!THIS IS AWESOME:
-Hello this is the first text file
+THIS IS AWESOME: 
+Hello this is the first text file,
+Hello this is the second text file!THIS IS AWESOME: 
+Hello this is the first text file,
 Hello this is the second text file!
