+++ index.js
@@ -1,1 +1,4 @@
-
+/*
+! STREAMS
+? There are 4 types of streams 1) Writeable , 2) Readable, 3) Duplex, 4) Transform.
+*/
